# Conflicts:
#	src/main/java/org/architecture_helper/gui/Main.java

+++ src/main/java/org/architecture_helper/backend/LRU.java
@@ -22,7 +22,7 @@ public class LRU {
     }
 
     public void pushToTop(int index) {
-        values.remove(index);
+        values.remove((Integer)index);
         values.add(0, index);
     }
 }
